+++ lib/src/_application/error/error_component.dart
@@ -1,4 +1,4 @@
-library plato.archives.components.application.error;
+library plato.archives.components.error;
 
 import 'package:angular/angular.dart';
 import 'package:angular_components/angular_components.dart';