+++ lib/src/_application/workflow/workflow_component.dart
@@ -1,0 +1,11 @@
+library plato.archives.components.workflow;
+
+import 'workflow_service.dart';
+
+/// The [WorkflowComponent] class...
+class WorkflowComponent {
+  final WorkflowService _workflowService;
+
+  /// The [WorkflowComponent] constructor...
+  WorkflowComponent (this._workflowService);
+}