+++ lib/src/archive/archive.dart
@@ -1,0 +1,11 @@
+library plato.archives.models.archive;
+
+/// The [Archive] class...
+class Archive {
+  final String id;
+
+  final String title;
+
+  /// The [Archive] constructor...
+  Archive (this.id, this.title);
+}