+++ lib/src/archive/archive_component.dart
@@ -1,0 +1,11 @@
+library plato.archives.components.archive;
+
+import 'archive.dart';
+
+/// The [ArchiveComponent] class...
+class ArchiveComponent {
+  Archive archive;
+
+  /// The [ArchiveComponent] constructor...
+  ArchiveComponent();
+}