+++ lib/src/archive/archive_selection_component.dart
@@ -1,0 +1,13 @@
+library plato.archives.components.archive.selection;
+
+import 'archive.dart';
+
+/// The [ArchiveSelectionComponent] class...
+class ArchiveSelectionComponent {
+  List<Archive> archives;
+
+  Archive selected;
+
+  /// The [ArchiveSelectionComponent] constructor...
+  ArchiveSelectionComponent();
+}